Cleaned up code to remove all warnings.

+++ src/kernel/basic_commands.c
@@ -27,7 +27,7 @@ void END(char *args) {
 void PAGE(char *args) {
     void* page = malloc(1000);
     kprint("Page: ");
-    kprintn(hex_to_ascii(page));
+    kprintn(hex_to_ascii((int)page));
 
     UNUSED(args);
 }
